Restructure qualification tests with subdirectory
Moved X-engine tests into a baseline_correlation_products subdirectory
(I'm not committed to the name, but it's what the SE docs call that
output).

This is mainly so that we can parametrise these tests difference to
antenna_channelised_voltage tests, where we don't need to test different
numbers of antennas.
